REFACTOR: separate out rendering/buffer logic
create tabpage components
