+++ lua/bufferline/tabpages.lua
@@ -1,4 +1,9 @@
+local ui = require("bufferline.ui")
+local config = require("bufferline.config")
 local constants = require("bufferline.constants")
+
+local fn = vim.fn
+local api = vim.api
 
 local M = {}
 
@@ -20,13 +25,12 @@ local function render(tabpage, is_active, style, highlights)
   return hl .. tab_click_component(tabpage.tabnr) .. name .. separator, length
 end
 
---- @param style string
---- @param prefs table
-function M.get(style, prefs)
+function M.get()
   local tabpages = {}
   local tabs = vim.fn.gettabinfo()
   local current_tab = vim.fn.tabpagenr()
-  local highlights = prefs.highlights
+  local highlights = config.get("highlights")
+  local style = config.get("options").separator_style
 
   -- use contiguous numbers to ensure contiguous keys in the table i.e. an array
   -- rather than an object
@@ -43,6 +47,47 @@ function M.get(style, prefs)
     }
   end
   return tabpages
+end
+
+local function get_tab_name(tab_num)
+  local no_name
+  if not api.nvim_tabpage_is_valid(tab_num) then
+    return "[No name]"
+  end
+  local buflist = fn.tabpagebuflist(tab_num)
+  -- tabpagebuflist can return 0 if the tab page number was invalid
+  -- if this happens show no name
+  if buflist == 0 or #buflist < 1 then
+    return no_name
+  end
+  local winnr = fn.tabpagewinnr(tab_num)
+  return fn.bufname(buflist[winnr]), buflist[winnr]
+end
+
+---@param state BufferlineState
+---@return Tabpage[]
+function M.get_components(state)
+  local options = config.get("options")
+  local tabs = api.nvim_list_tabpages()
+  local Tabpage = require("bufferline.models").Tabpage
+  ---@type Tabpage[]
+  local components = {}
+  for i, tab_num in ipairs(tabs) do
+    local path, buf_num = get_tab_name(tab_num)
+    components[i] = Tabpage:new({
+      path = path,
+      buf = buf_num,
+      id = tab_num,
+      ordinal = i,
+      diagnostics = {},
+      name_formatter = options.name_formatter,
+      hidden = false,
+      focusable = true,
+    })
+  end
+  return vim.tbl_map(function(tab)
+    return ui.element(state, tab)
+  end, components)
 end
 
 return M
